Rework high-level service-definition framework.
Instead of having to subclass the “SInterface” base class that was
defined with a metaclass, you now use the “@sservice”
decorator-creating function instead. This way you can define the
service class however you like, in particular being able to
instantiate it with whatever arguments (or no arguments) that you
need.