Use a callback to to call `_reaper` instead of scheduling it
asyncio is only maintaining weak references to task. To avoid them
being destroyed while still pending, we need to keep a reference on
them. Previously, a function was scheduled at each tick to cleanup
done tasks. Instead, we add a callback to the task to cleanup its own
reference when done.

Fix #44.